Moved resolve_text_background_xyxy() function into utils.py
